requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==2.0.6
django-import-export==2.5.0
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.8
platformdirs==2.3.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
typing-extensions==3.10.0.2
virtualenv==20.8.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
